+++ calculation.rb
@@ -1,10 +1,14 @@
-def ci_lower_bound(pos, n, confidence)
-  if n == 0
+require 'statistics2'
+
+#Wilson's confidence interval method of ranking. Uses current number of votes as a sample of the whole population.
+def ci_lower_bound(positive_ratings, total_ratings, confidence)
+  #The confidence argument tells the ranking algorithm how aggresive it can be with its assumptions
+  if total_ratings == 0
     return 0
   end
   z = Statistics2.pnormaldist(1-(1-confidence)/2)
-  phat = 1.0*pos/n
-  value = (phat + z*z/(2*n) - z * Math.sqrt((phat*(1-phat)+z*z/(4*n))/n))/(1+z*z/n)
+  phat = 1.0*positive_ratings/total_ratings
+  value = (phat + z*z/(2*total_ratings) - z * Math.sqrt((phat*(1-phat)+z*z/(4*total_ratings))/total_ratings))/(1+z*z/total_ratings)
   value*10**4
 end
 
@@ -38,6 +42,14 @@ def sum_total_ratings(ratings)
     end
   end
   return total_ratings
+end
+
+def my_ranker(positive_ratings,total_ratings,standard_deviation)
+  if positive_ratings > 0
+    return -1/(positive_ratings/60) + (positive_ratings/total_ratings)*2 - (standard_deviation)**2
+  else
+    return -1.0/0.0
+  end
 end
 
 def sum_score_method2(ratings)
@@ -74,13 +86,5 @@ def sd (ratings,total_ratings,positive_ratings)
         sq_diff += ratings[index] * ((-4-mean)**2)
     end
     return Math.sqrt(sq_diff/total_ratings)
-  end
-end
-
-def rank(positive_ratings,total_ratings,standard_deviation)
-  if positive_ratings > 0
-    return -1/(positive_ratings/60) + (positive_ratings/total_ratings)*2 - (standard_deviation)**2
-  else
-    return -1.0/0.0
   end
 end
