+++ scrape_restaurant_page.rb
@@ -1,7 +1,6 @@
 def scrape_restaurant_page(city,url)
   agent = Mechanize.new       #create new mechanize object
   #send GET request to restaurant url
-#Try until works
   page = agent.get(url, "User-Agent" => "Ruby/#{RUBY_VERSION}")  #some websites do not respond to scrapers that do not identify themselves
 
   name = page.search("h1#HEADING.heading_name").text.strip       #get name of restaurant with css selector
