--- test.rb
@@ -1,2 +1,0 @@
-require_relative 'db_query'
-db_query
